[Autodiff] Fail on non-static range in backwards (#229)
