+++ gstaichi/program/compile_config.h
@@ -50,6 +50,7 @@ struct CompileConfig {
   bool cpu_block_dim_adaptive;
   int default_gpu_block_dim;
   int gpu_max_reg;
+  bool ad_stack_experimental_enabled{false};
   int ad_stack_size{0};  // 0 = adaptive
   // The default size when the GsTaichi compiler is unable to automatically
   // determine the autodiff stack size.
