+++ gstaichi/python/export_lang.cpp
@@ -210,6 +210,8 @@ void export_lang(py::module &m) {
       .def_readwrite("fast_math", &CompileConfig::fast_math)
       .def_readwrite("advanced_optimization",
                      &CompileConfig::advanced_optimization)
+      .def_readwrite("ad_stack_experimental_enabled",
+                     &CompileConfig::ad_stack_experimental_enabled)
       .def_readwrite("ad_stack_size", &CompileConfig::ad_stack_size)
       .def_readwrite("flatten_if", &CompileConfig::flatten_if)
       .def_readwrite("make_thread_local", &CompileConfig::make_thread_local)
