+++ python/gstaichi/lang/ast/ast_transformer.py
@@ -42,6 +42,8 @@ from gstaichi.lang.struct import Struct, StructType
 from gstaichi.types import primitive_types
 from gstaichi.types.utils import is_integral
 
+AutodiffMode = _ti_core.AutodiffMode
+
 
 def reshape_list(flat_list: list[Any], target_shape: Sequence[int]) -> list[Any]:
     if len(target_shape) < 2:
@@ -940,7 +942,9 @@ class ASTTransformer(Builder):
 
             for_di = _ti_core.DebugInfo(ctx.get_pos_info(node))
             ctx.ast_builder.begin_frontend_range_for(loop_var.ptr, begin.ptr, end.ptr, for_di)
+            ctx.loop_depth += 1
             build_stmts(ctx, node.body)
+            ctx.loop_depth -= 1
             ctx.ast_builder.end_frontend_range_for()
         return None
 
@@ -983,7 +987,9 @@ class ASTTransformer(Builder):
                 )
                 if i + 1 < len(targets):
                     I._assign(I - target_tmp * ndrange_var.acc_dimensions[i + 1])
+            ctx.loop_depth += 1
             build_stmts(ctx, node.body)
+            ctx.loop_depth -= 1
             ctx.ast_builder.end_frontend_range_for()
         return None
 
@@ -1017,7 +1023,9 @@ class ASTTransformer(Builder):
                 impl.subscript(ctx.ast_builder, target_var, i)._assign(target_tmp + ndrange_var.bounds[i][0])
                 if i + 1 < len(ndrange_var.dimensions):
                     I._assign(I - target_tmp * ndrange_var.acc_dimensions[i + 1])
+            ctx.loop_depth += 1
             build_stmts(ctx, node.body)
+            ctx.loop_depth -= 1
             ctx.ast_builder.end_frontend_range_for()
         return None
 
@@ -1051,7 +1059,9 @@ class ASTTransformer(Builder):
                 loop_var = node.iter.ptr
                 expr_group = expr.make_expr_group(*_vars)
                 impl.begin_frontend_struct_for(ctx.ast_builder, expr_group, loop_var)
+                ctx.loop_depth += 1
                 build_stmts(ctx, node.body)
+                ctx.loop_depth -= 1
                 ctx.ast_builder.end_frontend_struct_for()
         return None
 
@@ -1074,7 +1084,9 @@ class ASTTransformer(Builder):
                 node.iter.ptr._type,
                 _ti_core.DebugInfo(impl.get_runtime().get_current_src_info()),
             )
+            ctx.loop_depth += 1
             build_stmts(ctx, node.body)
+            ctx.loop_depth -= 1
             ctx.mesh = None
             ctx.ast_builder.end_frontend_mesh_for()
         return None
@@ -1105,7 +1117,9 @@ class ASTTransformer(Builder):
             entry_expr.type_check(impl.get_runtime().prog.config())
             mesh_idx = mesh.MeshElementFieldProxy(ctx.mesh, node.iter.ptr.to_element_type, entry_expr)
             ctx.create_variable(target, mesh_idx)
+            ctx.loop_depth += 1
             build_stmts(ctx, node.body)
+            ctx.loop_depth -= 1
             ctx.ast_builder.end_frontend_range_for()
 
         return None
@@ -1143,6 +1157,8 @@ class ASTTransformer(Builder):
                 and isinstance(node.iter.func, ast.Name)
                 and node.iter.func.id == "range"
             ):
+                if ctx.loop_depth > 0 and ctx.autodiff_mode == AutodiffMode.REVERSE and not ctx.adstack_enabled:
+                    raise exception.GsTaichiCompilationError("Cannot use non static range in Backwards mode")
                 return ASTTransformer.build_range_for(ctx, node)
             elif isinstance(node.iter, ast.IfExp):
                 # Handle inline if expression as the top level iterator expression, e.g.:
