+++ python/gstaichi/lang/ast/ast_transformer_utils.py
@@ -8,6 +8,7 @@ from enum import Enum
 from textwrap import TextWrapper
 from typing import TYPE_CHECKING, Any, List
 
+from gstaichi._lib import core as _ti_core
 from gstaichi._lib.core.gstaichi_python import ASTBuilder
 from gstaichi.lang import impl
 from gstaichi.lang._ndrange import ndrange
@@ -18,6 +19,9 @@ from gstaichi.lang.exception import (
     GsTaichiSyntaxError,
     handle_exception_from_cpp,
 )
+
+AutodiffMode = _ti_core.AutodiffMode
+
 
 if TYPE_CHECKING:
     from gstaichi.lang.kernel_impl import (
@@ -189,7 +193,10 @@ class ASTTransformerContext:
         start_lineno: int,
         ast_builder: ASTBuilder | None,
         is_real_function: bool,
+        autodiff_mode: AutodiffMode,
     ):
+        from gstaichi import extension  # pylint: disable=import-outside-toplevel
+
         self.func = func
         self.local_scopes: list[dict[str, Any]] = []
         self.loop_scopes: List[LoopScopeAttribute] = []
@@ -222,6 +229,16 @@ class ASTTransformerContext:
         self.is_real_function = is_real_function
         self.kernel_args: list = []
         self.only_parse_function_def: bool = False
+        self.autodiff_mode = autodiff_mode
+        self.loop_depth: int = 0
+
+        self.adstack_enabled: bool = (
+            _ti_core.is_extension_supported(
+                impl.current_cfg().arch,
+                extension.adstack,
+            )
+            and impl.current_cfg().ad_stack_experimental_enabled
+        )
 
     # e.g.: FunctionDef, Module, Global
     def variable_scope_guard(self):
