+++ python/gstaichi/lang/kernel_impl.py
@@ -299,6 +299,8 @@ def _get_tree_and_ctx(
     assert current_kernel is not None
     current_kernel.visited_functions.add(function_source_info)
 
+    autodiff_mode = current_kernel.autodiff_mode
+
     gstaichi_callable = current_kernel.gstaichi_callable
     is_pure = gstaichi_callable is not None and gstaichi_callable.is_pure
     global_vars = _get_global_vars(self.func)
@@ -328,6 +330,7 @@ def _get_tree_and_ctx(
         file=function_source_info.filepath,
         ast_builder=ast_builder,
         is_real_function=is_real_function,
+        autodiff_mode=autodiff_mode,
     )
