+++ tests/python/test_ad_for.py
@@ -3,7 +3,7 @@ import gstaichi as ti
 from tests import test_utils
 
 
-@test_utils.test()
+@test_utils.test(require=ti.extension.adstack)
 def test_ad_nested_for():
     N = 5
